+++ actix-limitation/src/middleware.rs
@@ -6,12 +6,14 @@ use actix_web::{
     body::EitherBody,
     cookie::Cookie,
     dev::{forward_ready, Service, ServiceRequest, ServiceResponse, Transform},
-    http::header::COOKIE,
+    http::{header::COOKIE, StatusCode},
     web, Error, HttpResponse,
 };
 
 use crate::Limiter;
 
+/// Rate limit middleware.
+#[derive(Debug)]
 pub struct RateLimiter;
 
 impl<S, B> Transform<S, ServiceRequest> for RateLimiter
@@ -22,8 +24,8 @@ where
 {
     type Response = ServiceResponse<EitherBody<B>>;
     type Error = Error;
-    type InitError = ();
     type Transform = RateLimiterMiddleware<S>;
+    type InitError = ();
     type Future = Ready<Result<Self::Transform, Self::InitError>>;
 
     fn new_transform(&self, service: S) -> Self::Future {
@@ -33,6 +35,8 @@ where
     }
 }
 
+/// Rate limit middleware service.
+#[derive(Debug)]
 pub struct RateLimiterMiddleware<S> {
     service: Rc<S>,
 }
@@ -57,10 +61,10 @@ where
             .expect("web::Data<Limiter> should be set in app data for RateLimiter middleware")
             .clone();
 
-        let forbidden = HttpResponse::Forbidden().finish().map_into_right_body();
         let (key, fallback) = key(&req, limiter.clone());
 
         let service = Rc::clone(&self.service);
+
         let key = match key {
             Some(key) => key,
             None => match fallback {
@@ -76,12 +80,15 @@ where
             },
         };
 
-        let service = Rc::clone(&self.service);
         Box::pin(async move {
             let status = limiter.count(key.to_string()).await;
+
             if status.is_err() {
-                warn!("403. Rate limit exceed error for {}", key);
-                Ok(req.into_response(forbidden))
+                log::warn!("Rate limit exceed error for {}", key);
+
+                Ok(req.into_response(
+                    HttpResponse::new(StatusCode::TOO_MANY_REQUESTS).map_into_right_body(),
+                ))
             } else {
                 service
                     .call(req)
@@ -98,7 +105,7 @@ fn key(req: &ServiceRequest, limiter: web::Data<Limiter>) -> (Option<String>, Op
     let cookies = req.headers().get_all(COOKIE);
     let cookie = cookies
         .filter_map(|i| i.to_str().ok())
-        .find(|i| i.contains(&limiter.cookie_name));
+        .find(|i| i.contains(limiter.cookie_name.as_ref()));
 
     let fallback = match cookie {
         Some(value) => Cookie::parse(value).ok().map(|i| i.to_string()),